Added Basic PDF creation

+++ src/main/java/it/unitn/disi/wp/cup/persistence/dao/jdbc/JDBCExamDAO.java
@@ -41,7 +41,7 @@ public class JDBCExamDAO extends JDBCDAO<Exam, Long> implements ExamDAO {
 
             exam.setId(rs.getLong("id"));
             exam.setName(rs.getString("name"));
-            exam.setPrice(rs.getFloat("price"));
+            exam.setPrice(rs.getShort("price"));
         } catch (SQLException ex) {
             throw new DAOException("Impossible to set Exam by ResultSet", ex);
         }
