+++ src/main/java/it/unitn/disi/wp/cup/persistence/dao/jdbc/JDBCMedicineDAO.java
@@ -41,7 +41,7 @@ public class JDBCMedicineDAO extends JDBCDAO<Medicine, Long> implements Medicine
 
             medicine.setId(rs.getLong("id"));
             medicine.setName(rs.getString("name"));
-            medicine.setPrice(rs.getFloat("price"));
+            medicine.setPrice(rs.getShort("price"));
         } catch (SQLException ex) {
             throw new DAOException("Impossible to set Medicine by ResultSet", ex);
         }
